refactor: Cli directory structure
