+++ crates/testa-cli/src/commands/check.rs
@@ -1,0 +1,18 @@
+use std::{error::Error, path::PathBuf};
+
+use crate::compiler::{compile_file, parse_file};
+
+pub fn check_command(
+    files: Vec<PathBuf>,
+    syntax_only: bool,
+    warnings: bool,
+) -> Result<(), Box<dyn Error>> {
+    for file in files {
+        if syntax_only {
+            parse_file(&file)?;
+        } else {
+            compile_file(&file, warnings)?;
+        }
+    }
+    Ok(())
+}
