+++ crates/testa-cli/src/commands/mod.rs
@@ -1,0 +1,5 @@
+pub mod check;
+pub mod generate;
+pub mod info;
+pub mod init;
+pub mod lsp;
