+++ crates/testa-cli/src/templates/advanced.testa
@@ -1,0 +1,21 @@
+@output csv;
+@output_path "./advanced.csv";
+
+enum Status {
+    Active => 70;
+    Inactive => 20;
+    Pending => 10;
+}
+
+template User {
+    id = int;
+    username = string;
+    status = Status;
+}
+
+template Admin : User {
+    permissions = [string][count=1..=5];
+    override status = "Active";
+}
+
+@generate Admin [3]
