+++ crates/testa-cli/src/templates/main.testa
@@ -1,0 +1,11 @@
+@output csv;
+@output_path "./output.csv";
+
+template User {
+    id = int [range=1..=1000];
+    name = string;
+    email = string;
+    active = bool;
+}
+
+@generate User [10];
